Refactor decoder into workspace library
