+++ README.md
@@ -1,62 +1,58 @@
 # subtitle-fast
 
-An experimental Rust workspace for streaming Y-plane data from decoded H.264 frames. The crate exposes a trait-based
-interface so multiple decoding backends can coexist behind a unified asynchronous stream.
+A Rust workspace for streaming Y-plane data from decoded H.264 frames. The workspace currently contains:
+
+- `subtitle-fast-decoder`: a library crate that exposes the `YPlaneStreamProvider` trait alongside several native decoding
+  backends.
+- `subtitle-fast`: a CLI crate that wires the decoder into a binary for manual backend testing.
 
 ## Architecture Overview
 
 - `core`: Shared primitives, including the `YPlaneStreamProvider` trait, `YPlaneFrame` metadata, and error handling.
 - `config`: Runtime configuration, enabling backend selection via environment variables or direct API usage.
-- `backends`: Concrete backend modules. The FFmpeg implementation decodes frames in a blocking task and forwards Y-plane
-  data through a Tokio `mpsc` channel. Mock and placeholder providers are included for environments lacking native
-  libraries.
+- `backends`: Concrete backend modules. FFmpeg, VideoToolbox, OpenH264, and GStreamer all feed Y-plane data into an async
+  Tokio stream.
 
-Backends are toggled through Cargo features:
+Backends are toggled through Cargo features on the library crate:
 
 | Feature | Description |
 | ------- | ----------- |
-| `backend-mock` (default) | Generates synthetic frames for testing and local development. |
 | `backend-ffmpeg` | Uses `ffmpeg-next` bindings for pure-Rust decoding without invoking the CLI binary. |
-| `backend-videotoolbox` | Placeholder for a future macOS VideoToolbox backend. |
-| `backend-openh264` | Placeholder for an OpenH264-based decoder with multi-threaded support. |
-| `backend-gstreamer` | Placeholder for a GStreamer pipeline using appsink. |
+| `backend-videotoolbox` | Hardware-accelerated decoding on macOS using VideoToolbox via Objective-C bindings. |
+| `backend-openh264` | Multi-threaded decoding using the OpenH264 Rust bindings. |
+| `backend-gstreamer` | Streams decoded I420 frames from a GStreamer pipeline. |
 
-## Running
+## Running the CLI
 
-1. Ensure the desired backend feature is enabled. By default the mock backend runs, which does not require external
-   dependencies.
-2. Optionally configure runtime selection using environment variables:
-   - `SUBFAST_BACKEND` &mdash; choose a backend (`mock`, `ffmpeg`, `videotoolbox`, `openh264`, `gstreamer`).
-   - `SUBFAST_INPUT` &mdash; path to an H.264/MP4 asset when using FFmpeg or other file-based backends.
-3. Execute the binary:
+1. Enable the desired backend features on the workspace member. The CLI forwards feature flags to the decoder crate.
+2. Provide the input asset path as the first positional argument. Backend selection can be overridden with
+   `SUBFAST_BACKEND`, and file paths can be supplied via `SUBFAST_INPUT` when invoking the library directly.
+3. Run the binary:
 
 ```bash
-cargo run --release
+cargo run --release -- <video-path>
 ```
 
-To enable the FFmpeg backend:
+To run with a specific backend (for example, FFmpeg):
 
 ```bash
 SUBFAST_BACKEND=ffmpeg \
-SUBFAST_INPUT=/path/to/video.mp4 \
-cargo run --release --no-default-features --features backend-ffmpeg
+cargo run --release --features backend-ffmpeg -- <video-path>
 ```
 
 ## Testing
 
-Run the standard tests with the default (mock) backend:
+Integration tests live under the decoder crate and require backend-specific assets. For example, to exercise the FFmpeg
+backend:
 
 ```bash
-cargo test
+SUBFAST_TEST_ASSET=/path/to/video.mp4 \
+cargo test -p subtitle-fast-decoder --features backend-ffmpeg
 ```
 
-To run FFmpeg integration tests, provide a valid test asset:
-
-```bash
-SUBFAST_TEST_ASSET=/path/to/video.mp4 cargo test --no-default-features --features backend-ffmpeg
-```
+Backends that rely on platform-specific APIs (e.g., VideoToolbox) will only run on compatible targets.
 
 ## Continuous Integration
 
-See `.github/workflows/ci.yml` for a minimal GitHub Actions pipeline that exercises formatting and tests with the default
+See `.github/workflows/ci.yml` for a minimal GitHub Actions pipeline that exercises formatting and tests with a selected
 feature set.
