+++ TODO.md
@@ -33,7 +33,7 @@
 ## Cross-Cutting Concerns
 - [x] Define configuration layer selecting backend at runtime or compile time.
 - [ ] Implement unit/integration tests using sample H.264 assets for each backend (where supported). *(FFmpeg, OpenH264, and VideoToolbox covered; GStreamer pending)*
-- [x] Provide mock/test double for Y-plane provider for environments lacking native support.
+- [ ] Provide mock/test double for Y-plane provider for environments lacking native support. *(Removed during workspace split; reconsider lightweight stub if needed.)*
 - [x] Document usage, backend selection, and platform requirements in README.
 - [x] Set up CI to build with minimal backend (e.g., OpenH264) and run lint/tests.
