+++ crates/subtitle-fast-decoder/src/backends/mod.rs
@@ -4,9 +4,6 @@ pub mod ffmpeg;
 #[cfg(feature = "backend-gstreamer")]
 pub mod gstreamer;
 
-#[cfg(feature = "backend-mock")]
-pub mod mock;
-
 #[cfg(feature = "backend-openh264")]
 pub mod openh264;
