+++ crates/subtitle-fast-decoder/tests/ffmpeg_stream.rs
@@ -1,29 +1,28 @@
-#![cfg(all(feature = "backend-videotoolbox", target_os = "macos"))]
+#![cfg(feature = "backend-ffmpeg")]
 
 use std::env;
 use std::path::PathBuf;
 
-use subtitle_fast::{Backend, Configuration, YPlaneStreamProvider};
+use subtitle_fast_decoder::{Backend, Configuration, YPlaneStreamProvider};
 use tokio_stream::StreamExt;
 
 #[tokio::test(flavor = "multi_thread")]
-async fn videotoolbox_backend_streams_frames() {
+async fn ffmpeg_backend_requires_asset() {
     let asset = match env::var("SUBFAST_TEST_ASSET") {
         Ok(value) => PathBuf::from(value),
         Err(_) => {
-            eprintln!("skipping videotoolbox backend test - SUBFAST_TEST_ASSET not set");
+            eprintln!("skipping ffmpeg backend test - SUBFAST_TEST_ASSET not set");
             return;
         }
     };
 
     let mut config = Configuration::default();
-    config.backend = Backend::VideoToolbox;
+    config.backend = Backend::Ffmpeg;
     config.input = Some(asset);
-
     let provider = match config.create_provider() {
         Ok(provider) => provider,
         Err(err) => {
-            panic!("failed to initialize VideoToolbox backend: {err:?}");
+            panic!("failed to initialize ffmpeg backend: {err:?}");
         }
     };
 
@@ -31,7 +30,7 @@ async fn videotoolbox_backend_streams_frames() {
     let frame = stream
         .next()
         .await
-        .expect("videotoolbox backend should produce at least one frame");
+        .expect("ffmpeg backend should produce at least one frame");
     let frame = frame.expect("frame decoding should succeed");
     assert!(frame.width() > 0);
     assert!(frame.height() > 0);
