+++ crates/subtitle-fast-decoder/tests/openh264_stream.rs
@@ -3,7 +3,7 @@
 use std::env;
 use std::path::PathBuf;
 
-use subtitle_fast::{Backend, Configuration, YPlaneStreamProvider};
+use subtitle_fast_decoder::{Backend, Configuration, YPlaneStreamProvider};
 use tokio_stream::StreamExt;
 
 #[tokio::test(flavor = "multi_thread")]
