+++ crates/subtitle-fast/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "subtitle-fast"
+version = "0.1.0"
+edition = "2024"
+
+[features]
+default = []
+backend-ffmpeg = ["subtitle-fast-decoder/backend-ffmpeg"]
+backend-videotoolbox = ["subtitle-fast-decoder/backend-videotoolbox"]
+backend-openh264 = ["subtitle-fast-decoder/backend-openh264"]
+backend-gstreamer = ["subtitle-fast-decoder/backend-gstreamer"]
+
+[dependencies]
+subtitle-fast-decoder = { path = "../subtitle-fast-decoder", default-features = false }
+tokio = { version = "1", features = ["macros", "rt-multi-thread"] }
+tokio-stream = "0.1"
